// src/shared/api/mocks/schema/generated.ts
export interface components {
    schemas: {
        Story: {
            id: string;
            name: string;
            description: string;
            img: string;
            author: {
                id: string;
                login: string;
            };
            tags: string[];
            /** Format: int32 */
            sceneCount: number;
            /** Format: int32 */
            passes: number;
            /** Format: int32 */
            likes: number;
            isLiked: boolean;
            /** Format: date */
            date: string;
        };
        StoryLike: {
            id?: string;
            storyId?: string;
            userId?: string;
        };
        Scene: {
            /** @description Scene ID */
            id: string;
            /** @description ID of the story this scene belongs to */
            storyId: string;
            /**
             * Format: int32
             * @description Scene number within the story
             */
            number: number;
            /**
             * @description Type of the scene ('default', 'end')
             * @enum {string}
             */
            type: "default" | "end";
            /** @description Possible answers/choices for this scene */
            answers: components["schemas"]["Answer"][];
            /** @description Title of the scene */
            title: string;
            /** @description Detailed description of the scene */
            description: string;
            /**
             * Format: url
             * @description URL of the image associated with the scene
             */
            img: string;
            /**
             * Format: int32
             * @description Number of times this scene has been passed
             */
            passes?: number;
        };
        Answer: {
            /** @description ID of the answer choice */
            id: string;
            /** @description The text displayed for this answer choice */
            text: string;
            /** @description ID of the next scene, or null if this is an ending choice */
            nextSceneNumber: number;
        };
        User: {
            id: string;
            login: string;
        };
        LoginRequest: {
            login: string;
            /** Format: password */
            password: string;
        };
        Error: {
            message: string;
            code: string;
        };
        StoryLikeDTO: {
            isLiked?: boolean;
        };
        Session: {
            id: string;
            login: string;
        };
        AuthResponse: {
            session?: components["schemas"]["Session"];
        };
    };
    responses: {
        /** @description Bad request */
        BadRequestError: {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["Error"];
            };
        };
        /** @description Unauthorized */
        UnauthorizedError: {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["Error"];
            };
        };
        /** @description Resource not found */
        NotFoundError: {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["Error"];
            };
        };
    };
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}